Update konversation to 1.7.6
Summary:
Konversation v1.7.6 is a small translation update & bugfix release fixing the build with Qt >= 5.13 and polishing the configuration dialog a bit.

- Fixed loading of nick icon theme to deal with multiple copies
- Fixed configuration dialogs to have consistent indentation of subordinate options as well as spacings between labels & fields
- Fixed building against Qt 5.15.
- Fixed building against Qt 5.13.

**Package changes**:
- Enabled LTO

Test Plan:
I can still access to the Solus IRC channels and write on them.
LTO adds ~40 seconds of linking time with the benefit of ~400 KiB saved from the executable file.

Reviewers: #triage_team, JoshStrobl

Reviewed By: #triage_team, JoshStrobl

Subscribers: JoshStrobl

Differential Revision: https://dev.getsol.us/D9810